\name{spatialRtBayes-package}
\alias{spatialRtBayes-package}
\alias{spatialRtBayes}
\docType{package}
\title{
\packageTitle{spatialRtBayes}
}
\description{
\packageDescription{spatialRtBayes}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{spatialRtBayes}

Maintainer: \packageMaintainer{spatialRtBayes}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = spatialRtBayes)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{spatialRtBayes}}
%\section{Documentation Index}{\packageIndices{spatialRtBayes}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
